Split TestHost into separate files
