+++ ktor-hosts/ktor-test-host/src/org/jetbrains/ktor/testing/TestApplicationResponse.kt
@@ -1,0 +1,51 @@
+package org.jetbrains.ktor.testing
+
+import org.jetbrains.ktor.application.*
+import org.jetbrains.ktor.cio.*
+import org.jetbrains.ktor.host.*
+import org.jetbrains.ktor.http.*
+import org.jetbrains.ktor.response.*
+import org.jetbrains.ktor.util.*
+
+class TestApplicationResponse(call: ApplicationCall, respondPipeline: ApplicationResponsePipeline = ApplicationResponsePipeline()) : BaseApplicationResponse(call, respondPipeline) {
+    internal val realContent = lazy { ByteBufferWriteChannel() }
+
+    @Volatile
+    private var closed = false
+
+    override fun setStatus(statusCode: HttpStatusCode) {
+    }
+
+    override val headers: ResponseHeaders = object : ResponseHeaders() {
+        private val headersMap = ValuesMapBuilder(true)
+        private val headers: ValuesMap by lazy { headersMap.build() }
+
+        override fun hostAppendHeader(name: String, value: String) {
+            if (closed)
+                throw UnsupportedOperationException("Headers can no longer be set because response was already completed")
+            headersMap.append(name, value)
+        }
+
+        override fun getHostHeaderNames(): List<String> = headers.names().toList()
+        override fun getHostHeaderValues(name: String): List<String> = headers.getAll(name).orEmpty()
+    }
+
+
+    val content: String?
+        get() = if (realContent.isInitialized()) {
+            realContent.value.toString(headers[HttpHeaders.ContentType]?.let { ContentType.parse(it).charset() } ?: Charsets.UTF_8)
+        } else {
+            null
+        }
+
+    val byteContent: ByteArray?
+        get() = if (realContent.isInitialized()) {
+            realContent.value.toByteArray()
+        } else {
+            null
+        }
+
+    fun close() {
+        closed = true
+    }
+}
